Nothing is working, need to push
Connections more or less work, however, not all the functionality for them is in there yet, and they seem to not be deleted from the .asset file. Need to organize it more, currently working on small window UI. Also do note if you get an exception and the graph refuses to respond, just delete it and make a new one.